fastapi_server/requirements.txt
# FastAPI e dependências principais
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Banco de dados
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Autenticação e segurança
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validação e serialização
pydantic==2.5.0
pydantic-settings==2.1.0

# CORS e middleware
fastapi-cors==0.0.6

# Utilitários
python-dotenv==1.0.0
email-validator==2.1.0

# Desenvolvimento e testes
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Documentação
mkdocs==1.5.3
mkdocs-material==9.4.8